fixed the issues present in the pull request
