+++ backend/src/main/java/com/bookadaisical/IUserService.java
@@ -1,0 +1,9 @@
+package com.bookadaisical;
+
+import java.util.List;
+
+public interface IUserService {
+    
+    List<User> getAllUsers();
+
+}
