+++ backend/src/main/java/com/bookadaisical/User.java
@@ -104,3 +104,4 @@ public class User {
     }
     
 }
+
