+++ backend/src/main/java/com/bookadaisical/UserService.java
@@ -2,8 +2,21 @@ package com.bookadaisical;
 
 import java.util.List;
 
-public interface UserService {
-    
-    List<User> getAllUsers();
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
+@Service
+public class UserService implements IUserService {
+    
+    private final UserRepository userRepository;
+
+    @Autowired
+    public UserService(UserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
+
+    @Override
+    public List<User> getAllUsers() {
+        return userRepository.findAll();
+    }
 }
